refact: replicator test
* refactored the class, and kept the custom conflict in separate class
* kept the replicatorTest alone in with only helper class.
* replicatorTest_Main & replicatorTest_CustomConflict are the new classes.
